+++ cmd/health-monitor_test.go
@@ -35,7 +35,6 @@ package cmd
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"testing"
 
@@ -141,7 +140,7 @@ func (suite *hmonTestSuite) TestHmonInvalidConfigFile() {
 func (suite *hmonTestSuite) TestHmonWithConfigFailure() {
 	defer suite.cleanupTest()
 
-	confFile, err := ioutil.TempFile("", "conf*.yaml")
+	confFile, err := os.CreateTemp("", "conf*.yaml")
 	suite.assert.Nil(err)
 	cfgFileHmonTest := confFile.Name()
 	defer os.Remove(cfgFileHmonTest)