+++ cmd/mount_test.go
@@ -36,7 +36,6 @@ package cmd
 import (
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
@@ -131,7 +130,7 @@ func (suite *mountTestSuite) TestMountDirNotExists() {
 func (suite *mountTestSuite) TestMountDirNotEmpty() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	tempDir := filepath.Join(mntDir, "tempdir")
 
@@ -164,7 +163,7 @@ func (suite *mountTestSuite) TestMountPathNotProvided() {
 func (suite *mountTestSuite) TestUnsupportedConfigFileType() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -178,7 +177,7 @@ func (suite *mountTestSuite) TestUnsupportedConfigFileType() {
 func (suite *mountTestSuite) TestConfigFileNotFound() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -197,7 +196,7 @@ func (suite *mountTestSuite) TestConfigFileNotFound() {
 func (suite *mountTestSuite) TestConfigFileNotProvided() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -209,7 +208,7 @@ func (suite *mountTestSuite) TestConfigFileNotProvided() {
 func (suite *mountTestSuite) TestDefaultConfigFile() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -242,7 +241,7 @@ func (suite *mountTestSuite) TestDefaultConfigFile() {
 func (suite *mountTestSuite) TestInvalidLogLevel() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -254,7 +253,7 @@ func (suite *mountTestSuite) TestInvalidLogLevel() {
 func (suite *mountTestSuite) TestCliParamsV1() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -270,7 +269,7 @@ func (suite *mountTestSuite) TestCliParamsV1() {
 func (suite *mountTestSuite) TestStreamAttrCacheOptionsV1() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -287,7 +286,7 @@ func (suite *mountTestSuite) TestStreamAttrCacheOptionsV1() {
 func (suite *mountTestSuite) TestInvalidLibfuseOption() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -303,7 +302,7 @@ func (suite *mountTestSuite) TestInvalidLibfuseOption() {
 func (suite *mountTestSuite) TestUndefinedLibfuseOption() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -319,7 +318,7 @@ func (suite *mountTestSuite) TestUndefinedLibfuseOption() {
 func (suite *mountTestSuite) TestInvalidUmaskValue() {
 	defer suite.cleanupTest()
 
-	mntDir, err := ioutil.TempDir("", "mntdir")
+	mntDir, err := os.MkdirTemp("", "mntdir")
 	suite.assert.Nil(err)
 	defer os.RemoveAll(mntDir)
 
@@ -433,7 +432,7 @@ func (suite *mountTestSuite) TestMountOptionVaildate() {
 }
 
 func TestMountCommand(t *testing.T) {
-	confFile, err := ioutil.TempFile("", "conf*.yaml")
+	confFile, err := os.CreateTemp("", "conf*.yaml")
 	if err != nil {
 		t.Error("Failed to create config file")
 	}