+++ component/azstorage/block_blob.go
@@ -1192,7 +1192,8 @@ func (bb *BlockBlob) stageAndCommitModifiedBlocks(name string, data []byte, offs
 		bb.blobAccCond,
 		bb.Config.defaultTier,
 		nil, // datalake doesn't support tags here
-		bb.downloadOptions.ClientProvidedKeyOptions)
+		bb.downloadOptions.ClientProvidedKeyOptions,
+		azblob.ImmutabilityPolicyOptions{})
 	if err != nil {
 		log.Err("BlockBlob::stageAndCommitModifiedBlocks : Failed to commit block list to blob %s [%s]", name, err.Error())
 		return err
@@ -1241,7 +1242,8 @@ func (bb *BlockBlob) StageAndCommit(name string, bol *common.BlockOffsetList) er
 			// azblob.BlobAccessConditions{ModifiedAccessConditions: azblob.ModifiedAccessConditions{IfMatch: bol.Etag}},
 			bb.Config.defaultTier,
 			nil, // datalake doesn't support tags here
-			bb.downloadOptions.ClientProvidedKeyOptions)
+			bb.downloadOptions.ClientProvidedKeyOptions,
+			azblob.ImmutabilityPolicyOptions{})
 		if err != nil {
 			log.Err("BlockBlob::StageAndCommit : Failed to commit block list to blob %s [%s]", name, err.Error())
 			return err