+++ test/e2e_tests/data_validation_test.go
@@ -1,3 +1,4 @@
+//go:build !unittest
 // +build !unittest
 
 /*
@@ -38,7 +39,6 @@ package e2e_tests
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"os/exec"
@@ -139,7 +139,7 @@ func (suite *dataValidationTestSuite) TestSmallFileData() {
 	srcFile.Close()
 
 	// write to file in the local directory
-	err = ioutil.WriteFile(localFilePath, minBuff, 0777)
+	err = os.WriteFile(localFilePath, minBuff, 0777)
 	suite.Equal(nil, err)
 
 	suite.copyToMountDir(localFilePath, remoteFilePath)
@@ -168,7 +168,7 @@ func (suite *dataValidationTestSuite) TestMediumFileData() {
 	srcFile.Close()
 
 	// write to file in the local directory
-	err = ioutil.WriteFile(localFilePath, medBuff, 0777)
+	err = os.WriteFile(localFilePath, medBuff, 0777)
 	suite.Equal(nil, err)
 
 	suite.copyToMountDir(localFilePath, remoteFilePath)
@@ -197,7 +197,7 @@ func (suite *dataValidationTestSuite) TestLargeFileData() {
 	srcFile.Close()
 
 	// write to file in the local directory
-	err = ioutil.WriteFile(localFilePath, largeBuff, 0777)
+	err = os.WriteFile(localFilePath, largeBuff, 0777)
 	suite.Equal(nil, err)
 
 	suite.copyToMountDir(localFilePath, remoteFilePath)
@@ -222,7 +222,7 @@ func (suite *dataValidationTestSuite) TestDataValidationNegative() {
 	srcFile.Close()
 
 	// write to file in the local directory
-	err = ioutil.WriteFile(localFilePath, minBuff, 0777)
+	err = os.WriteFile(localFilePath, minBuff, 0777)
 	suite.Equal(nil, err)
 
 	// copy local file to mounted directory
@@ -263,17 +263,17 @@ func validateMultipleFilesData(jobs <-chan int, results chan<- string, fileSize 
 
 		// write to file in the local directory
 		if fileSize == "huge" {
-			err = ioutil.WriteFile(localFilePath, hugeBuff, 0777)
+			err = os.WriteFile(localFilePath, hugeBuff, 0777)
 		} else if fileSize == "large" {
 			if strings.ToLower(quickTest) == "true" {
-				err = ioutil.WriteFile(localFilePath, hugeBuff, 0777)
+				err = os.WriteFile(localFilePath, hugeBuff, 0777)
 			} else {
-				err = ioutil.WriteFile(localFilePath, largeBuff, 0777)
+				err = os.WriteFile(localFilePath, largeBuff, 0777)
 			}
 		} else if fileSize == "medium" {
-			err = ioutil.WriteFile(localFilePath, medBuff, 0777)
+			err = os.WriteFile(localFilePath, medBuff, 0777)
 		} else {
-			err = ioutil.WriteFile(localFilePath, minBuff, 0777)
+			err = os.WriteFile(localFilePath, minBuff, 0777)
 		}
 		suite.Equal(nil, err)