+++ test/e2e_tests/file_test.go
@@ -1,3 +1,4 @@
+//go:build !unittest
 // +build !unittest
 
 /*
@@ -39,7 +40,6 @@ import (
 	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"strconv"
@@ -140,7 +140,7 @@ func (suite *fileTestSuite) TestFileCreatSpclChar() {
 	_, err = os.Stat(fileName)
 	suite.Equal(nil, err)
 
-	files, err := ioutil.ReadDir(suite.testPath)
+	files, err := os.ReadDir(suite.testPath)
 	suite.Equal(nil, err)
 	suite.GreaterOrEqual(len(files), 1)
 
@@ -168,7 +168,7 @@ func (suite *fileTestSuite) TestFileCreatEncodeChar() {
 	_, err = os.Stat(fileName)
 	suite.Equal(nil, err)
 
-	files, err := ioutil.ReadDir(suite.testPath)
+	files, err := os.ReadDir(suite.testPath)
 	suite.Equal(nil, err)
 	suite.GreaterOrEqual(len(files), 1)
 
@@ -211,7 +211,7 @@ func (suite *fileTestSuite) TestFileCreateMultiSpclCharWithinSpclDir() {
 	_, err = os.Stat(fileName)
 	suite.Equal(nil, err)
 
-	files, err := ioutil.ReadDir(speclDirName)
+	files, err := os.ReadDir(speclDirName)
 	suite.Equal(nil, err)
 	suite.GreaterOrEqual(len(files), 1)
 
@@ -271,7 +271,7 @@ func (suite *fileTestSuite) TestFileWriteSmall() {
 	suite.Equal(nil, err)
 	srcFile.Close()
 
-	err = ioutil.WriteFile(fileName, suite.minBuff, 0777)
+	err = os.WriteFile(fileName, suite.minBuff, 0777)
 	suite.Equal(nil, err)
 
 	suite.fileTestCleanup([]string{fileName})
@@ -284,10 +284,10 @@ func (suite *fileTestSuite) TestFileReadSmall() {
 	suite.Equal(nil, err)
 	srcFile.Close()
 
-	err = ioutil.WriteFile(fileName, suite.minBuff, 0777)
+	err = os.WriteFile(fileName, suite.minBuff, 0777)
 	suite.Equal(nil, err)
 
-	data, err := ioutil.ReadFile(fileName)
+	data, err := os.ReadFile(fileName)
 	suite.Equal(nil, err)
 	suite.Equal(len(data), len(suite.minBuff))
 
@@ -318,7 +318,7 @@ func (suite *fileTestSuite) TestFileTruncate() {
 	err = os.Truncate(fileName, 2)
 	suite.Equal(nil, err)
 
-	data, err := ioutil.ReadFile(fileName)
+	data, err := os.ReadFile(fileName)
 	suite.Equal(nil, err)
 	suite.LessOrEqual(2, len(data))
 
@@ -414,7 +414,7 @@ func (suite *fileTestSuite) TestFileCreateMulti() {
 	fileName := dirName + "/multi"
 	for i := 0; i < 10; i++ {
 		newFile := fileName + strconv.Itoa(i)
-		err := ioutil.WriteFile(newFile, suite.medBuff, 0777)
+		err := os.WriteFile(newFile, suite.medBuff, 0777)
 		suite.Equal(nil, err)
 	}
 	suite.fileTestCleanup([]string{dirName})
@@ -439,7 +439,7 @@ func (suite *fileTestSuite) TestLinkCreate() {
 	suite.Equal(nil, err)
 	f.Close()
 	symName := suite.testPath + "/small.lnk"
-	err = ioutil.WriteFile(fileName, suite.minBuff, 0777)
+	err = os.WriteFile(fileName, suite.minBuff, 0777)
 	suite.Equal(nil, err)
 
 	err = os.Symlink(fileName, symName)
@@ -460,9 +460,9 @@ func (suite *fileTestSuite) TestLinkRead() {
 	err = os.Symlink(fileName, symName)
 	suite.Equal(nil, err)
 
-	err = ioutil.WriteFile(fileName, suite.minBuff, 0777)
+	err = os.WriteFile(fileName, suite.minBuff, 0777)
 	suite.Equal(nil, err)
-	data, err := ioutil.ReadFile(fileName)
+	data, err := os.ReadFile(fileName)
 	suite.Equal(nil, err)
 	suite.Equal(len(data), len(suite.minBuff))
 	suite.fileTestCleanup([]string{fileName})
@@ -503,7 +503,7 @@ func (suite *fileTestSuite) TestLinkRenameTarget() {
 	err = os.Rename(fileName, fileNameNew)
 	suite.Equal(nil, err)
 
-	_, err = ioutil.ReadFile(symName)
+	_, err = os.ReadFile(symName)
 	// we expect that to fail
 	suite.NotEqual(nil, err)
 
@@ -525,13 +525,13 @@ func (suite *fileTestSuite) TestLinkDeleteReadTarget() {
 	f.Close()
 	err = os.Symlink(fileName, symName)
 	suite.Equal(nil, err)
-	err = ioutil.WriteFile(fileName, suite.minBuff, 0777)
+	err = os.WriteFile(fileName, suite.minBuff, 0777)
 	suite.Equal(nil, err)
 
 	err = os.Remove(symName)
 	suite.Equal(nil, err)
 
-	data, err := ioutil.ReadFile(fileName)
+	data, err := os.ReadFile(fileName)
 	suite.Equal(nil, err)
 	suite.Equal(len(data), len(suite.minBuff))