+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.4 (WIP)
+**Features**
+- Added new config parameter "max-fuse-threads" under "libfuse" config to control max threads allowed at libfuse layer.
+- Added new config parameter 'refresh-sec' in 'file-cache'. When file-cache-timeout is set to a large value, this field can control when to refresh the file if file in container has changed.
+
+**Bug Fixes**
+- [#1116](https://github.com/Azure/azure-storage-fuse/issues/1116)] Relative path for tmp-cache is resulting into file read-write failure.
+
 ## 2.0.3 (2023-04-26)
 **Bug Fixes**
 - [#1080](https://github.com/Azure/azure-storage-fuse/issues/1080) HNS rename flow does not encode source path correctly.